fix: price decimal type

+++ back-end/src/database/models/ProductModel.js
@@ -2,7 +2,7 @@ module.exports = (sequelize, DataTypes) => {
   const Product = sequelize.define('Product', {
     id: { type: DataTypes.INTEGER, allowNull: false, autoIncrement: true, primaryKey: true },
     name: { type: DataTypes.STRING, allowNull: false },
-    price: { type: DataTypes.INTEGER, allowNull: false },
+    price: { type: DataTypes.DECIMAL, allowNull: false },
     urlImage: { type: DataTypes.STRING, allowNull: false },
   }, {
     timestamps: false,
